BM-8

+++ src/main/java/com/finance/money/infrastructure/persistence/entities/Account.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Entity
@@ -22,7 +23,7 @@ public class Account {
     private String accountNumber;
 
     @Column(name = "BALANCE", nullable = false)
-    private Double accountBalance;
+    private BigDecimal accountBalance;
 
     @Column(name = "CUSTOMER_ID")
     @ManyToOne
